requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
Django==5.1.3
executing==2.1.0
fastapi==0.129.0
Flask==3.1.1
fonttools==4.54.1
h11==0.16.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
ipykernel==6.29.5
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
missingno==0.5.2
multimethod==1.12
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pandas-profiling==3.2.0
parso==0.8.4
phik==0.12.4
pillow==11.0.0
pip-review==1.3.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
py4j==0.10.9.7
pycryptodome==3.23.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyjokes==0.8.3
pyparsing==3.2.0
pyspark==3.5.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.8.0
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.52.1
tangled-up-in-unicode==0.2.0
threadpoolctl==3.6.0
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.41.0
visions==0.7.6
wcwidth==0.2.13
Werkzeug==3.1.3
